+++ sdk/resourcemanager/azure-resourcemanager-samples/src/main/java/com/azure/resourcemanager/servicebus/samples/ServiceBusPublishSubscribeBasic.java
@@ -19,7 +19,6 @@ import com.azure.resourcemanager.servicebus.models.AuthorizationKeys;
 import com.azure.resourcemanager.servicebus.models.KeyType;
 import com.azure.resourcemanager.servicebus.models.NamespaceAuthorizationRule;
 import com.azure.resourcemanager.servicebus.models.NamespaceSku;
-import com.azure.resourcemanager.servicebus.models.RegenerateAccessKeyParameters;
 import com.azure.resourcemanager.servicebus.models.ServiceBusNamespace;
 import com.azure.resourcemanager.servicebus.models.ServiceBusSubscription;
 import com.azure.resourcemanager.servicebus.models.Topic;
@@ -143,8 +142,7 @@ public final class ServiceBusPublishSubscribeBasic {
             AuthorizationKeys keys = namespaceAuthorizationRules.iterator().next().getKeys();
             Utils.print(keys);
             System.out.println("Regenerating secondary key for authorization rule ...");
-            keys = namespaceAuthorizationRules.iterator().next()
-                .regenerateKey(new RegenerateAccessKeyParameters().withKeyType(KeyType.SECONDARY_KEY));
+            keys = namespaceAuthorizationRules.iterator().next().regenerateKey(KeyType.SECONDARY_KEY);
             Utils.print(keys);
 
             //=============================================================