+++ sdk/resourcemanager/azure-resourcemanager-samples/src/main/java/com/azure/resourcemanager/servicebus/samples/ServiceBusQueueBasic.java
@@ -20,7 +20,6 @@ import com.azure.resourcemanager.servicebus.models.KeyType;
 import com.azure.resourcemanager.servicebus.models.NamespaceAuthorizationRule;
 import com.azure.resourcemanager.servicebus.models.NamespaceSku;
 import com.azure.resourcemanager.servicebus.models.Queue;
-import com.azure.resourcemanager.servicebus.models.RegenerateAccessKeyParameters;
 import com.azure.resourcemanager.servicebus.models.ServiceBusNamespace;
 
 /**
@@ -141,8 +140,7 @@ public final class ServiceBusQueueBasic {
             AuthorizationKeys keys = namespaceAuthorizationRules.iterator().next().getKeys();
             Utils.print(keys);
             System.out.println("Regenerating secondary key for authorization rule ...");
-            keys = namespaceAuthorizationRules.iterator().next()
-                .regenerateKey(new RegenerateAccessKeyParameters().withKeyType(KeyType.SECONDARY_KEY));
+            keys = namespaceAuthorizationRules.iterator().next().regenerateKey(KeyType.SECONDARY_KEY);
             Utils.print(keys);
 
             //=============================================================