+++ sdk/resourcemanager/azure-resourcemanager-servicebus/src/test/java/com/azure/resourcemanager/servicebus/ServiceBusOperationsTests.java
@@ -25,7 +25,6 @@ import com.azure.resourcemanager.servicebus.models.NamespaceAuthorizationRule;
 import com.azure.resourcemanager.servicebus.models.NamespaceSku;
 import com.azure.resourcemanager.servicebus.models.Queue;
 import com.azure.resourcemanager.servicebus.models.QueueAuthorizationRule;
-import com.azure.resourcemanager.servicebus.models.RegenerateAccessKeyParameters;
 import com.azure.resourcemanager.servicebus.models.ServiceBusNamespace;
 import com.azure.resourcemanager.servicebus.models.ServiceBusSubscription;
 import com.azure.resourcemanager.servicebus.models.Topic;
@@ -414,7 +413,7 @@ public class ServiceBusOperationsTests extends ResourceManagerTestBase {
         Assertions.assertNotNull(nsRuleKeys.secondaryKey());
         Assertions.assertNotNull(nsRuleKeys.primaryConnectionString());
         Assertions.assertNotNull(nsRuleKeys.secondaryConnectionString());
-        nsRuleKeys = foundNsRule.regenerateKey(new RegenerateAccessKeyParameters().withKeyType(KeyType.PRIMARY_KEY));
+        nsRuleKeys = foundNsRule.regenerateKey(KeyType.PRIMARY_KEY);
         if (!isPlaybackMode()) {
             Assertions.assertNotEquals(nsRuleKeys.primaryKey(), primaryKey);
         }